StyleTTS2/requirements.txt
SoundFile
torchaudio
munch
torch
pydub
pyyaml
librosa
nltk
matplotlib
accelerate
transformers
einops
einops-exts
tqdm
typing
typing-extensions
git+https://github.com/resemble-ai/monotonic_align.git
scipy
phonemizer
cached-path
gradio
gruut
#tortoise-tts
txtsplit